requirements.txt
asgiref==3.2.8
Django==3.0.7
djangorestframework==3.11.0
Faker==0.9.1
mixer==6.1.3
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
